26281: Fix handlingPreviewAllData.test.ts
